// src/features/charinfo/types.ts
// /src/features/charinfo/types.ts

export interface ItemData {
  item_name?: string;
  item_icon?: string;
  item_equipment_slot?: string;
  scroll_upgrade?: number;
  potential_option_grade?: string;
  potential_option_1?: string;
  potential_option_2?: string;
  potential_option_3?: string;
  additional_potential_option_grade?: string;
  additional_potential_option_1?: string;
  additional_potential_option_2?: string;
  additional_potential_option_3?: string;
}